style: move Pagination/WithPagination to component with index
