+++ front/app/advanced-search/components/CommunitiesTable.tsx
@@ -3,6 +3,7 @@
 import { PropsWithChildren } from 'react';
 
 import { AdvancedSearchCommunity } from '@/app/models/community';
+import { Pagination } from '@/components/Pagination';
 import { Button } from '@/components/ui/button';
 import {
   Table,
@@ -15,7 +16,6 @@ import {
 import { cn, formatNumber, stringifyCommunityType } from '@/utils/utils';
 import { ArrowUpDown } from 'lucide-react';
 
-import { Pagination } from '../../community/[siren]/components/Pagination/Pagination';
 import { AdvancedSearchOrder, useOrderParams } from '../hooks/useOrderParams';
 import { usePaginationParams } from '../hooks/usePaginationParams';
