+++ front/components/Pagination/index.tsx
@@ -1,0 +1,2 @@
+export * from './Pagination';
+export * from './WithPagination';
